split out strategies from back testing, add test case
